requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-extensions==2.1.9
django-rest-framework==0.1.0
djangorestframework==3.9.4
pkg-resources==0.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Werkzeug==0.15.4
